requirements.txt
# Core geospatial
geopandas>=0.14
rasterio>=1.3
shapely>=2.0
pyproj>=3.6
fiona>=1.9
parquet>=1.3.1

# Data
numpy>=1.24
pandas>=2.0

# Distance raster (Feature 2)
scipy>=1.11

# Planetary Computer STAC access
pystac-client>=0.7
planetary-computer>=1.0

# HTTP
requests>=2.31

# Progress
tqdm>=4.65

# Model / evaluation
torch>=2.0
segmentation-models-pytorch>=0.3
scikit-learn>=1.3
matplotlib>=3.7

# Visualization / mapping
folium>=0.14
Pillow>=10.0

# Optimizer (MILP solver)
pulp>=2.7
